backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
sendgrid
twilio
python-dotenv
pydantic
httpx
